requirement.txt
# This file may be used to install all required package using:
# $ pip install --requirement <this file> -yes
# platform: osx-64
baidu-aip==2.1.0.0
certifi==2017.11.5
chardet==3.0.4
idna==2.6
Pillow==5.0.0
pip==9.0.1
requests==2.18.4
setuptools==36.4.0
urllib3==1.22
wheel==0.29.0
zhon==1.1.5
